backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
rapidfuzz
numpy
scikit-learn
aiosqlite
pyjwt
bcrypt
python-multipart
httpx
